deployment_v15/unzipped_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase>=2.10.0
python-dotenv==1.0.0
boto3==1.28.63
pydantic==2.4.2
httpx>=0.26.0
python-multipart==0.0.6
